DEPRECATED/tools/requirements.txt
matplotlib
numpy
paho-mqtt
pandas
pytz
PyYaml
